Encodes fonts with base64 and compiles fonts.scss
Now all .woff files will be encoded with base64 and inlined into fonts.css. This reduces the amount of HTTP requests but increases the size of fonts.css.
Once loaded the whole fonts.css will be stored in localStorage, so the client doesn't need to request it again.

Adds gulp-base64 and gulp-rename as devDependencies.